move dialog options to layout
